fix(go-rewrite): dedupe newCanonicalStore test helper
Two test files independently declared newCanonicalStore in
package api_test with different signatures (1-arg vs 2-arg). Move
the canonical 1-arg variant to helpers_external_test.go; rename
the get_node_by_key 2-arg variant to newCanonicalStoreForTenant.

Refs #407
